move nav related constants to own file
